node engine: fix tier-3 low-severity swallowed errors
- inbound_node_ips.go: log the ignored IP-blob unmarshal (was silently dropping
  stored IPs); check the marshal error and skip the write instead of persisting
  an empty IP set.
- xray.go: log when subscription-outbound load fails (was silently dropping all
  sub outbounds from the generated config).
- node.go probe: distinguish "success=false" from "success with empty payload"
  instead of printing the wrong cause.
- node_traffic_sync_job.go: compute per-inbound speed in float to avoid an int64
  overflow on a huge first-after-gap delta (display value).

go build + go vet clean; service and job tests pass.

+++ internal/web/job/node_traffic_sync_job.go
@@ -282,8 +282,10 @@ func (j *NodeTrafficSyncJob) nodeInboundSpeed() []*xray.Traffic {
 			dDown = 0
 		}
 		elapsed := max(now-prev.at, nodeInboundSpeedWindowMs)
-		up := dUp * nodeInboundSpeedWindowMs / elapsed
-		down := dDown * nodeInboundSpeedWindowMs / elapsed
+		// Float math avoids int64 overflow when a first-after-gap delta is huge
+		// (dUp * window would wrap); this is a display speed, precision loss is moot.
+		up := int64(float64(dUp) * float64(nodeInboundSpeedWindowMs) / float64(elapsed))
+		down := int64(float64(dDown) * float64(nodeInboundSpeedWindowMs) / float64(elapsed))
 		if up > 0 || down > 0 {
 			deltas = append(deltas, &xray.Traffic{Tag: tag, IsInbound: true, Up: up, Down: down})
 		}
